+++ src/main/java/webserver/managers/IndexManager.java
@@ -1,4 +1,4 @@
-package webserver;
+package webserver.managers;
 
 import java.io.File;
 import java.util.regex.Matcher;
