+++ src/main/java/webserver/managers/LoginManager.java
@@ -1,4 +1,4 @@
-package webserver;
+package webserver.managers;
 
 import db.Database;
 import java.io.File;
@@ -9,6 +9,7 @@ import java.util.UUID;
 import model.User;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import webserver.Cookie;
 import webserver.request.HttpRequest;
 import webserver.response.HttpResponse;
 import webserver.response.HttpResponseBody;
