+++ src/main/java/webserver/managers/LogoutManager.java
@@ -1,4 +1,4 @@
-package webserver;
+package webserver.managers;
 
 import java.io.File;
 import webserver.request.HttpRequest;
