+++ src/main/java/webserver/managers/RegisterManager.java
@@ -1,4 +1,4 @@
-package webserver;
+package webserver.managers;
 
 import db.Database;
 import java.io.File;
