+++ src/test/java/webserver/LoginManagerTest.java
@@ -9,6 +9,7 @@ import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import webserver.managers.LoginManager;
 import webserver.request.HttpRequest;
 import webserver.response.HttpResponse;
