+++ src/test/java/webserver/RegisterManagerTest.java
@@ -9,6 +9,7 @@ import model.User;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import webserver.managers.RegisterManager;
 import webserver.request.HttpRequest;
 
 class RegisterManagerTest {
